requirements.txt
sqlglot>=18.0.0
graphviz>=0.20.1
sqlparse>=0.4.4
pydot>=1.4.2
flask>=2.3.0
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
python-docx==1.1.2
requests>=2.25.0
PyMySQL>=1.0.0
bcrypt>=4.0.0
flask-limiter>=3.5.0
cryptography>=41.0.0

# AI检测模块依赖
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0